backend/requirements.txt
flask >= 2.0, < 2.1
zeroc-ice >= 3.7.8, < 4.0
pydantic >= 1.10.4, < 1.11
pytest >= 7.2.1, <8
typeguard >= 2.13.3, < 2.14
flask-cors >= 3.0.10, < 3.1
pyjwt >= 2.6.0, < 2.7
pytest-cov >= 4.0.0, < 4.1
